Move Map Marker component to the Map component
